ml/requirements.txt
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
lightgbm==4.1.0
joblib==1.3.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
supabase==2.0.3

